refactor(utils): move shared utils to a separate library
moved shared utils to a separate library;
ran tests, confirmed that app works;
