+++ angular.json
@@ -435,6 +435,39 @@
           "style": "scss"
         }
       }
+    },
+    "shared-util": {
+      "projectType": "library",
+      "root": "libs/shared-util",
+      "sourceRoot": "libs/shared-util/src",
+      "prefix": "nx-ng-starter",
+      "architect": {
+        "lint": {
+          "builder": "@angular-eslint/builder:lint",
+          "options": {
+            "eslintConfig": "libs/shared-util/.eslintrc",
+            "tsConfig": [
+              "libs/shared-util/tsconfig.lib.json",
+              "libs/shared-util/tsconfig.spec.json"
+            ],
+            "exclude": ["**/node_modules/**", "!libs/shared-util/**/*"]
+          }
+        },
+        "test": {
+          "builder": "@nrwl/jest:jest",
+          "options": {
+            "jestConfig": "libs/shared-util/jest.config.js",
+            "tsConfig": "libs/shared-util/tsconfig.spec.json",
+            "passWithNoTests": true,
+            "setupFile": "libs/shared-util/src/test-setup.ts"
+          }
+        }
+      },
+      "schematics": {
+        "@nrwl/angular:component": {
+          "style": "scss"
+        }
+      }
     }
   },
   "schematics": {
