+++ apps/nx-ng-starter/src/app/components/app-index/home/app-index-home.component.ts
@@ -1,7 +1,7 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { UntilDestroy, untilDestroyed } from '@ngneat/until-destroy';
 import { MarkdownService } from '@nx-ng-starter/shared-core/data-access';
-import { ETIMEOUT } from '@nx-ng-starter/shared-core/util';
+import { ETIMEOUT } from '@nx-ng-starter/shared-util';
 import { timer } from 'rxjs';
 import { map } from 'rxjs/operators';
