+++ libs/mocks-core/src/lib/mocks-core.module.ts
@@ -12,10 +12,9 @@ import { RouterTestingModule } from '@angular/router/testing';
 import { NgxsFormPluginModule } from '@ngxs/form-plugin';
 import { NgxsModule } from '@ngxs/store';
 import { CustomMaterialModule } from '@nx-ng-starter/shared-core/ui';
-import { getWindow, WINDOW } from '@nx-ng-starter/shared-core/util';
+import { getDocument, getWindow, WINDOW } from '@nx-ng-starter/shared-util';
 
 import { SharedCoreModule } from '../../../shared-core/src/lib/shared-core.module';
-import { getDocument } from '../../../shared-core/src/lib/util/general-purpose/factories';
 import { httpHandlersProviders, testingEnvironment } from './data-access';
 import { DummyComponent } from './ui/components/dummy.component.mock';
 import { dialogRefMockProvider } from './util/refs/dialog-ref.mock';
