+++ libs/shared-core/src/lib/data-access/http-handlers/http-handlers.service.ts
@@ -4,6 +4,7 @@ import { Inject, Injectable } from '@angular/core';
 import { TranslateService } from '@ngx-translate/core';
 import { HttpProgressService } from '@nx-ng-starter/shared-store/state/http-progress/http-progress.service';
 import { UserService } from '@nx-ng-starter/shared-store/state/user/user.service';
+import { EHTTP_STATUS, WINDOW } from '@nx-ng-starter/shared-util';
 import { HttpLink, HttpLinkHandler } from 'apollo-angular-link-http';
 import { ApolloLink, ExecutionResult, split } from 'apollo-link';
 import { ErrorResponse, onError } from 'apollo-link-error';
@@ -14,8 +15,6 @@ import memo from 'memo-decorator';
 import { MonoTypeOperatorFunction, Observable, throwError } from 'rxjs';
 import { catchError, take, tap, timeout } from 'rxjs/operators';
 
-import { WINDOW } from '../../util/general-purpose';
-import { EHTTP_STATUS } from '../../util/http/http-statuses.interface';
 import { APP_ENV, WebEnvironment } from '../interfaces';
 import { ToasterService } from '../toaster/toaster.service';
