--- libs/shared-core/src/lib/util/index.ts
@@ -1,8 +1,0 @@
-/**
- * Shared Core Library util barrel.
- */
-export * from './http/http-statuses.interface';
-export * from './http/message.interface';
-export * from './interfaces/index';
-export * from './general-purpose/index';
-export * from './ngxs/index';
