+++ libs/shared-store/src/lib/state/http-api/http-api.interface.ts
@@ -1,4 +1,4 @@
-import { IActionPayload, Message } from '@nx-ng-starter/shared-core/util';
+import { IActionPayload, Message } from '@nx-ng-starter/shared-util';
 import { Observable } from 'rxjs';
 
 export interface IHttpApiStateModel {
