+++ libs/shared-store/src/lib/state/user/user.interface.ts
@@ -1,5 +1,5 @@
 import { AppUser } from '@nx-ng-starter/shared-core/data-access';
-import { IActionPayload } from '@nx-ng-starter/shared-core/util';
+import { IActionPayload } from '@nx-ng-starter/shared-util';
 import { Observable } from 'rxjs';
 
 export class UserStateModel implements AppUser {
