+++ libs/shared-util/README.md
@@ -1,0 +1,7 @@
+# shared-util
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test shared-util` to execute the unit tests.
