+++ libs/shared-util/src/lib/general-purpose/generators.ts
