+++ libs/shared-util/src/lib/general-purpose/index.ts
