+++ libs/shared-util/src/lib/interfaces/counters.ts
