+++ libs/shared-util/src/lib/interfaces/index.ts
