+++ libs/shared-util/src/lib/interfaces/timeouts.ts
