+++ libs/shared-util/src/lib/ngxs/ngxs-action-payload.interface.ts
