+++ libs/shared-util/src/test-setup.ts
@@ -1,0 +1,4 @@
+import 'jest-preset-angular';
+import { setupJestJsdomGlobalMocks } from '@nx-ng-starter/mocks-core';
+
+setupJestJsdomGlobalMocks();
