+++ tools/shell/module-aliases.sh
@@ -20,6 +20,7 @@ MODULE_ALIAS_LIB_API_INTERFACE="lib:api-interface"
 MODULE_ALIAS_LIB_MOCKS_CORE="lib:mocks-core"
 MODULE_ALIAS_LIB_SHARED_CORE="lib:shared-core"
 MODULE_ALIAS_LIB_SHARED_STORE="lib:shared-store"
+MODULE_ALIAS_LIB_SHARED_UTIL="lib:shared-util"
 MODULE_ALIAS_LIB_PROTO="lib:proto"
 
 MODULE_ALIAS_NAMES=(
@@ -30,6 +31,7 @@ MODULE_ALIAS_NAMES=(
   MODULE_ALIAS_LIB_MOCKS_CORE
   MODULE_ALIAS_LIB_SHARED_CORE
   MODULE_ALIAS_LIB_SHARED_STORE
+  MODULE_ALIAS_LIB_SHARED_UTIL
   MODULE_ALIAS_LIB_PROTO
 )
 
@@ -41,6 +43,7 @@ MODULE_ALIAS_VARS=(
   "$MODULE_ALIAS_LIB_MOCKS_CORE"
   "$MODULE_ALIAS_LIB_SHARED_CORE"
   "$MODULE_ALIAS_LIB_SHARED_STORE"
+  "$MODULE_ALIAS_LIB_SHARED_UTIL"
   "$MODULE_ALIAS_LIB_PROTO"
 )
 
@@ -51,6 +54,7 @@ MODULE_ALIAS_NAMES_UNIT=(
   MODULE_ALIAS_LIB_MOCKS_CORE
   MODULE_ALIAS_LIB_SHARED_CORE
   MODULE_ALIAS_LIB_SHARED_STORE
+  MODULE_ALIAS_LIB_SHARED_UTIL
   MODULE_ALIAS_LIB_PROTO
 )
 
@@ -61,6 +65,7 @@ MODULE_ALIAS_VARS_UNIT=(
   "$MODULE_ALIAS_LIB_MOCKS_CORE"
   "$MODULE_ALIAS_LIB_SHARED_CORE"
   "$MODULE_ALIAS_LIB_SHARED_STORE"
+  "$MODULE_ALIAS_LIB_SHARED_UTIL"
   "$MODULE_ALIAS_LIB_PROTO"
 )
