+++ tsconfig.base.json
@@ -37,9 +37,9 @@
       "@nx-ng-starter/shared-core": ["libs/shared-core/src/index.ts"],
       "@nx-ng-starter/shared-core/data-access": ["libs/shared-core/src/lib/data-access/index.ts"],
       "@nx-ng-starter/shared-core/ui": ["libs/shared-core/src/lib/ui/index.ts"],
-      "@nx-ng-starter/shared-core/util": ["libs/shared-core/src/lib/util/index.ts"],
       "@nx-ng-starter/shared-store": ["libs/shared-store/src/index.ts"],
-      "@nx-ng-starter/shared-store/*": ["libs/shared-store/src/lib/*"]
+      "@nx-ng-starter/shared-store/*": ["libs/shared-store/src/lib/*"],
+      "@nx-ng-starter/shared-util": ["libs/shared-util/src/index.ts"]
     },
     "rootDir": ".",
     "skipDefaultLibCheck": true,
