+++ tsconfig.json
@@ -45,6 +45,12 @@
     },
     {
       "path": "./libs/shared-store/tsconfig.spec.json"
+    },
+    {
+      "path": "./libs/shared-util/tsconfig.lib.json"
+    },
+    {
+      "path": "./libs/shared-util/tsconfig.spec.json"
     }
   ]
 }
